.github/scripts/requirements.txt
httpx==0.27.2
PyGithub==2.3.0
tiktoken==0.7.0
huggingface_hub==0.26.2
hf_transfer==0.1.8
pyyaml==6.0.2
